renamed module holding log conversion utilites
