+++ cancat/scripts/pcap2cancat.py
@@ -3,7 +3,7 @@
 
 import sys
 import argparse
-import cancat.utils.parse
+import cancat.utils.convert
 
 
 def main():
@@ -16,4 +16,4 @@ def main():
     parser.add_argument('output', help='output cancat session')
     args = parser.parse_args(argv)
 
-    cancat.utils.parse.pcap2cancat(args.pcap, args.output)
+    cancat.utils.convert.pcap2cancat(args.pcap, args.output)
